Fall back to user type for parameters
Refactor ReverseEngineerRunner to make it smaller

fixes #869
