+++ src/GUI/RevEng.Core.50/RevEng.Core.50.csproj
@@ -28,6 +28,7 @@
     <Compile Include="..\RevEng.Core\ReverseEngineerOptionsExtensions.cs" Link="ReverseEngineerOptionsExtensions.cs" />
     <Compile Include="..\RevEng.Core\ReverseEngineerResultExtensions.cs" Link="ReverseEngineerResultExtensions.cs" />
     <Compile Include="..\RevEng.Core\ReverseEngineerRunner.cs" Link="ReverseEngineerRunner.cs" />
+    <Compile Include="..\RevEng.Core\ReverseEngineerScaffolder.cs" Link="ReverseEngineerScaffolder.cs" />
     <Compile Include="..\RevEng.Core\ServiceProviderBuilder.cs" Link="ServiceProviderBuilder.cs" />
     <Compile Include="..\RevEng.Core\SharedTypeExtensions.cs" Link="SharedTypeExtensions.cs" />
     <Compile Include="..\RevEng.Core\SqlServerHelper.cs" Link="SqlServerHelper.cs" />
